requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai>=0.8.0
requests>=2.32.3
cloudinary>=1.41.0
supabase>=2.7.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
httpx>=0.27.0
Pillow>=10.4.0
mangum==0.17.0
groq>=0.11.0
